fix: coins and underlying coins api

+++ contracts/interfaces/ICurve.sol
@@ -35,7 +35,7 @@ interface ICurveFi is IERC20 {
 
     function minter() external view returns (address);
 
-    function coins(int128) external view returns (address);
+    function coins(uint256) external view returns (address);
 
     function underlying_coins(int128) external view returns (address);
